refactor(patina): split the dynamic `is` suppression into its own module
The source-length check fails closed on an over-limit file that grows, and
this PR pushed both `no_unused_components.rs` (364 -> 459) and the shared
`linter/tests/sfc.rs` (566 -> 730) further past the 350-line limit.

Move the dynamic-`is` detection into `no_unused_components/dynamic_is.rs`,
the rule's unit tests into `no_unused_components/tests.rs`, and the
`vue/no-unused-components` lint tests out of the catch-all SFC test file into
`linter/tests/no_unused_components.rs`. Both files now sit below their base
line counts and no behavior changes.
