+++ crates/vize_patina/src/rules/vue/no_unused_components.rs
@@ -28,45 +28,24 @@
 
 #![allow(clippy::disallowed_macros)]
 
+mod dynamic_is;
+#[cfg(test)]
+mod tests;
+
+use self::dynamic_is::has_dynamic_is_binding;
 use crate::context::LintContext;
 use crate::diagnostic::Severity;
 use crate::rule::{Rule, RuleCategory, RuleMeta};
 use oxc_allocator::Allocator;
-use oxc_ast::ast::{
-    Expression, IdentifierReference, ImportDeclaration, ImportDeclarationSpecifier, TSType,
-};
+use oxc_ast::ast::{IdentifierReference, ImportDeclaration, ImportDeclarationSpecifier, TSType};
 use oxc_ast_visit::{Visit, walk::walk_ts_type};
-use oxc_parser::{ParseOptions, Parser};
+use oxc_parser::Parser;
 use oxc_span::SourceType;
 use vize_carton::{CompactString, FxHashSet, String, ToCompactString};
 use vize_croquis::naming::{is_pascal_case, to_pascal_case};
 use vize_croquis::{Croquis, ScopeData};
 use vize_relief::BindingType;
-use vize_relief::{ExpressionNode, PropNode, RootNode, TemplateChildNode};
-
-/// Whether a directive expression statically names a component.
-///
-/// The expression is parsed rather than inspected by its delimiters: matching
-/// quotes get `` `${name}` `` wrong in one direction (interpolated, so the name
-/// is only known at runtime) and `('MyPanel')` wrong in the other (a literal
-/// wearing parentheses).
-fn names_a_component(content: &str) -> bool {
-    let allocator = Allocator::default();
-    let Ok(expression) = Parser::new(&allocator, content, SourceType::ts())
-        .with_options(ParseOptions {
-            preserve_parens: false,
-            ..ParseOptions::default()
-        })
-        .parse_expression()
-    else {
-        return false;
-    };
-    match expression {
-        Expression::StringLiteral(_) => true,
-        Expression::TemplateLiteral(template) => template.expressions.is_empty(),
-        _ => false,
-    }
-}
+use vize_relief::RootNode;
 
 static META: RuleMeta = RuleMeta {
     name: "vue/no-unused-components",
@@ -140,72 +119,6 @@ impl NoUnusedComponents {
         names
     }
 
-    /// Whether the template binds `is` to something other than a string
-    /// literal, anywhere.
-    ///
-    /// `eslint-plugin-vue`'s `vue/no-unused-components` defaults
-    /// `ignoreWhenBindingPresent` to `true` and stops reporting for the whole
-    /// file when it sees one, because a dynamic `<component :is="resolved">` can
-    /// render any registered component and the rule cannot tell which. Reporting
-    /// anyway is how a legitimately-used component gets called unused — the
-    /// shape is common enough in real code to dominate this rule's output
-    /// (#3223).
-    ///
-    /// A literal (`:is="'MyPanel'"`) is exempt: it names its component, so the
-    /// registration is still checkable. A static `is="MyPanel"` attribute is not
-    /// a binding at all and does not suppress anything.
-    ///
-    /// The walk is an explicit worklist rather than recursion: the template
-    /// parser accepts up to `MAX_ELEMENT_NESTING_DEPTH` nested elements, and a
-    /// stack overflow on a valid file would take the LSP process down with it.
-    fn has_dynamic_is_binding<'a>(nodes: &'a [TemplateChildNode<'a>]) -> bool {
-        let mut worklist: Vec<&'a [TemplateChildNode<'a>]> = vec![nodes];
-        while let Some(children) = worklist.pop() {
-            for node in children {
-                match node {
-                    TemplateChildNode::Element(element) => {
-                        if element.props.iter().any(Self::is_dynamic_is_prop) {
-                            return true;
-                        }
-                        worklist.push(&element.children);
-                    }
-                    TemplateChildNode::If(node) => worklist.extend(
-                        node.branches
-                            .iter()
-                            .map(|branch| branch.children.as_slice()),
-                    ),
-                    TemplateChildNode::IfBranch(branch) => worklist.push(&branch.children),
-                    TemplateChildNode::For(node) => worklist.push(&node.children),
-                    _ => {}
-                }
-            }
-        }
-        false
-    }
-
-    fn is_dynamic_is_prop(prop: &PropNode<'_>) -> bool {
-        let PropNode::Directive(directive) = prop else {
-            return false;
-        };
-        if directive.name.as_str() != "bind" {
-            return false;
-        }
-        // A dynamic argument (`:[name]="x"`) can resolve to `is`, so it counts.
-        let Some(ExpressionNode::Simple(argument)) = directive.arg.as_ref() else {
-            return true;
-        };
-        if !argument.is_static {
-            return true;
-        }
-        if argument.content.as_str() != "is" {
-            return false;
-        }
-        !matches!(
-            directive.exp.as_ref(),
-            Some(ExpressionNode::Simple(expression)) if names_a_component(expression.content.as_str())
-        )
-    }
-
     fn component_name_matches(used: &str, registered: &str) -> bool {
         used == registered
             || vize_croquis::naming::names_match(used, registered)
@@ -231,7 +144,7 @@ impl Rule for NoUnusedComponents {
         if !ctx.has_analysis() {
             return;
         }
-        if Self::has_dynamic_is_binding(&root.children) {
+        if has_dynamic_is_binding(&root.children) {
             return;
         }
 
@@ -435,25 +348,5 @@ impl<'a> Visit<'a> for ScriptSetupComponentImportVisitor {
 impl ScriptSetupComponentImportVisitor {
     fn is_shadowed(&self, name: &str) -> bool {
         self.scopes.iter().rev().any(|scope| scope.contains(name))
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::NoUnusedComponents;
-    use crate::rule::{Rule, RuleCategory};
-
-    #[test]
-    fn test_meta() {
-        let rule = NoUnusedComponents::default();
-        assert_eq!(rule.meta().name, "vue/no-unused-components");
-        assert_eq!(rule.meta().category, RuleCategory::Essential);
-    }
-
-    #[test]
-    fn test_should_ignore() {
-        let rule = NoUnusedComponents::default();
-        assert!(rule.should_ignore("_Internal"));
-        assert!(!rule.should_ignore("MyComponent"));
     }
 }
